add deployment dockerfile

+++ scripts/build_clone_blogs.sh
@@ -1,13 +1,17 @@
-pushd src/main/content
+CUR_DIR="$(cd $(dirname $0) && pwd)"
+pushd "$CUR_DIR/../src/main/content"
 
-# For development purposes, lets always delete previously created folders
-# so that you can run this script to refresh your blog files
-rm -rf _posts _drafts
-rm -rf img/blog
-
-echo "Start cloning blogs repository..."
-
-git clone git@github.ibm.com:ICP4APPs/blogs.git --branch master blogs_temp
+# Allow user to pass in git url and branch, for example ./build_clone_blogs git@github.ibm.com:ICP4APPs/blogs.git master
+# Also allow the option to not pass in git url and branch, this is useful if the content already exists (Jenkins will have cloned the blogs before this step)
+if [ -n "$1" ] && [ -n "$2" ]; then
+	# For development purposes, lets always delete previously created folders
+	# so that you can run this script to refresh your blog files
+	rm -rf _posts _drafts
+	rm -rf img/blog
+	
+	echo "Start cloning blogs repository..."
+	git clone "$1" --branch "$2" blogs_temp
+fi
 
 mv blogs_temp/drafts/ .
 mv drafts/ _drafts
@@ -18,4 +22,4 @@ mv blogs_temp/img/blog img
 
 rm -rf blogs_temp
 popd
-echo "Done cloning blogs repository!"🚫
+echo "Done handling blogs repository!"
